+++ src/Tag/Textarea.php
@@ -4,8 +4,12 @@ declare(strict_types=1);
 
 namespace Yiisoft\Html\Tag;
 
+use Stringable;
 use Yiisoft\Html\Tag\Base\NormalTag;
 use Yiisoft\Html\Tag\Base\TagContentTrait;
+
+use function implode;
+use function is_array;
 
 /**
  * @link https://www.w3.org/TR/html52/sec-forms.html#the-textarea-element
@@ -35,9 +39,15 @@ final class Textarea extends NormalTag
         return $new;
     }
 
-    public function value(?string $value): self
+    /**
+     * @param string|string[]|Stringable|null $value
+     */
+    public function value(string|Stringable|array|null $value): self
     {
-        return $this->content($value ?? '');
+        $content = is_array($value)
+            ? implode("\n", $value)
+            : (string) $value;
+        return $this->content($content);
     }
 
     /**